requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
cryptography==3.3.1
Django==3.1.4
django-bootstrap-modal-forms==2.0.1
django-extensions==3.1.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
Pillow==8.0.1
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
